requirements.txt
streamlit
langchain
faiss-cpu
transformers
torch
sentence-transformers
nemo-guardrails
elasticsearch
PyYAML